+++ client/src/hooks/index.ts
@@ -11,3 +11,7 @@ export * from './useOtherUserdata';
 export * from './useToast';
 export * from './useSelectVideo';
 export * from './useSocket';
+export * from './useScroll';
+export * from './useChatSocket';
+export * from './useChats';
+export * from './useChatInfinite';
